attendly/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
black==23.9.1
click==8.1.7
decorator==5.1.1
Django==4.2.6
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
executing==2.0.0
flake8==6.1.0
gunicorn==21.2.0
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
wcwidth==0.2.8
wheel==0.41.2
